+++ _benchmarks/expressjs-throng-sessions/package.json
@@ -10,6 +10,8 @@
   "license": "ISC",
   "dependencies": {
     "express": "^4.16.0",
-    "express-session": "^1.15.6"
+    "express-session": "^1.15.6",
+    "session-memory-store": "^0.2.2",
+    "throng": "4.0.0"
   }
 }
