+++ _benchmarks/iris/main.go
@@ -11,7 +11,5 @@ func main() {
 		ctx.WriteString("value")
 	})
 
-	// 24 August 2017: Iris has a built'n version updater but we don't need it
-	// when benchmarking...
 	app.Run(iris.Addr(":5000"), iris.WithoutVersionChecker)
 }
