+++ _benchmarks/netcore-mvc-templates/Views/Home/Contact.cshtml
@@ -1,7 +1,4 @@
-﻿@{
-    ViewData["Title"] = "Contact";
-}
-<h2>@ViewData["Title"]</h2>
+﻿<h2>@ViewData["Title"]</h2>
 <h3>@ViewData["Message"]</h3>
 
 <address>
