+++ _benchmarks/netcore-mvc/netcore-mvc.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Folder Include="wwwroot\" />
+    <Folder Include="wwwroot/" />
   </ItemGroup>
 
   <ItemGroup>
